# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "signal-trck",
    "version": "0.1.0"
  },
  "paths": {
...
    "/pairs": {
      "get": {
        "summary": "List Pairs",
        "operationId": "list_pairs_pairs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PairResponse"
                  },
                  "type": "array",
                  "title": "Response List Pairs Pairs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Pair",
        "operationId": "create_pair_pairs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pairs/{pair_id}": {
      "delete": {
        "summary": "Remove Pair",
        "operationId": "remove_pair_pairs__pair_id__delete",
        "parameters": [
          {
            "name": "pair_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pair Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pairs/{pair_id}/candles": {
      "get": {
        "summary": "Get Candles",
        "operationId": "get_candles_pairs__pair_id__candles_get",
        "parameters": [
          {
            "name": "pair_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pair Id"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "1h",
                "1d",
                "1w"
              ],
              "type": "string",
              "default": "1d",
              "title": "Interval"
            }
          },
          {
            "name": "window_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3650,
              "minimum": 1,
              "default": 90,
              "title": "Window Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandleResponse"
                  },
                  "title": "Response Get Candles Pairs  Pair Id  Candles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pairs/{pair_id}/indicators/{name}": {
      "get": {
        "summary": "Get Indicator",
        "description": "Return a (potentially multi-output) indicator series.\n\nPer Decision 20: the ``name`` path param dispatches to a per-indicator\nparams shape. SMA/EMA/RSI take ``period``; MACD takes ``fast``/``slow``/\n``signal``; BB takes ``period``/``stddev``. Unknown ``name`` \u2192 422.",
        "operationId": "get_indicator_pairs__pair_id__indicators__name__get",
        "parameters": [
          {
            "name": "pair_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pair Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "1h",
                "1d",
                "1w"
              ],
              "type": "string",
              "default": "1d",
              "title": "Interval"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 500,
                  "minimum": 2
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          },
          {
            "name": "fast",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 200,
                  "minimum": 2
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fast"
            }
          },
          {
            "name": "slow",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 500,
                  "minimum": 3
                },
                {
                  "type": "null"
                }
              ],
              "title": "Slow"
            }
          },
          {
            "name": "signal",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 200,
                  "minimum": 2
                },
                {
                  "type": "null"
                }
              ],
              "title": "Signal"
            }
          },
          {
            "name": "stddev",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 10,
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stddev"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndicatorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pairs/{pair_id}/sr-candidates": {
      "get": {
        "summary": "Get Sr Candidates",
        "operationId": "get_sr_candidates_pairs__pair_id__sr_candidates_get",
        "parameters": [
          {
            "name": "pair_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pair Id"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "1h",
                "1d",
                "1w"
              ],
              "type": "string",
              "default": "1d",
              "title": "Interval"
            }
          },
          {
            "name": "window_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3650,
              "minimum": 1,
              "default": 180,
              "title": "Window Days"
            }
          },
          {
            "name": "top_n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Top N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SRCandidatesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pairs/{pair_id}/refresh": {
      "post": {
        "summary": "Refresh Pair",
        "description": "Fetch the most recent ``days`` worth of candles via the source adapter.",
        "operationId": "refresh_pair_pairs__pair_id__refresh_post",
        "parameters": [
          {
            "name": "pair_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pair Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/charts": {
      "get": {
        "summary": "List Charts",
        "operationId": "list_charts_charts_get",
        "parameters": [
          {
            "name": "pair_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pair Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 1000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChartListItemResponse"
                  },
                  "title": "Response List Charts Charts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Chart",
        "description": "Create a new chart. 409 ``CHART_SLUG_CONFLICT`` if slug exists.",
        "operationId": "create_chart_charts_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Chart"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chart"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/charts/{slug}": {
      "get": {
        "summary": "Get Chart",
        "operationId": "get_chart_charts__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chart"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Chart",
        "description": "Edit-in-place. 404 ``CHART_NOT_FOUND`` if slug doesn't exist.",
        "operationId": "update_chart_charts__slug__put",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Chart"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chart"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Chart",
        "operationId": "delete_chart_charts__slug__delete",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandleResponse": {
        "properties": {
          "ts_utc": {
            "type": "integer",
            "title": "Ts Utc"
          },
          "open": {
            "type": "number",
            "title": "Open"
          },
          "high": {
            "type": "number",
            "title": "High"
          },
          "low": {
            "type": "number",
            "title": "Low"
          },
          "close": {
            "type": "number",
            "title": "Close"
          },
          "volume": {
            "type": "number",
            "title": "Volume"
          }
        },
        "type": "object",
        "required": [
          "ts_utc",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "CandleResponse"
      },
...
      "CreatePairRequest": {
        "properties": {
          "pair_id": {
            "type": "string",
            "title": "Pair Id",
            "description": "Canonical id, e.g. 'coinbase:BTC-USD'"
          }
        },
        "type": "object",
        "required": [
          "pair_id"
        ],
        "title": "CreatePairRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PairResponse": {
        "properties": {
          "pair_id": {
            "type": "string",
            "title": "Pair Id"
          },
          "base": {
            "type": "string",
            "title": "Base"
          },
          "quote": {
            "type": "string",
            "title": "Quote"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "added_at": {
            "type": "integer",
            "title": "Added At"
          },
          "last_viewed_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Viewed At"
          },
          "is_pinned": {
            "type": "boolean",
            "title": "Is Pinned"
          },
          "pinned_context_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned Context Path"
          }
        },
        "type": "object",
        "required": [
          "pair_id",
          "base",
          "quote",
          "source",
          "added_at",
          "is_pinned"
        ],
        "title": "PairResponse"
      },
...
    }
  }
}